Fix HealthCheckUtil/MetricsUtil's support for environments other than test and production (#31)
The logic is no longer environment-specific (the environment-specific signatures have been deprecated, and delegate to the default one).
In all cases, if there is a registry in the servlet context, it is used.  If not, a fallback registry is instantiated on first access and then used.
This supports all cases where the application is running with a properly initialized servlet context, as well as cases (such as testing)
where the servlet context is not available.